+++ modules/@apostrophecms/area/index.js
@@ -25,7 +25,6 @@ module.exports = {
     ];
     self.richTextWidgetTypes = [];
     self.widgetManagers = {};
-    self.widgetOperations = [];
 
     self.enableBrowserData();
     self.addDeduplicateWidgetIdsMigration();
@@ -705,21 +704,6 @@ module.exports = {
           }
         });
       },
-      addWidgetOperation(operation) {
-        if (!operation.name || !operation.label || !operation.modal) {
-          throw self.apos.error('invalid', 'addWidgetOperation requires name, label and modal properties.');
-        }
-
-        if (operation.secondaryLevel !== true && !operation.icon) {
-          throw self.apos.error('invalid', 'addWidgetOperation requires the icon property at primary level.');
-        }
-
-        self.widgetOperations = self.widgetOperations.filter(
-          ({ name }) => name !== operation.name
-        );
-
-        self.widgetOperations.push(operation);
-      },
       getBrowserData(req) {
         const widgets = {};
         const widgetEditors = {};
@@ -744,13 +728,6 @@ module.exports = {
           contextualWidgetDefaultData[name] = manager.options.defaultData || {};
         });
 
-        const widgetOperations = self.widgetOperations.filter(({ permission }) => {
-          if (permission?.action && permission?.type) {
-            return self.apos.permission.can(req, permission.action, permission.type, permission.mode || 'draft');
-          }
-          return true;
-        });
-
         return {
           components: {
             editor: 'AposAreaEditor',
@@ -763,8 +740,7 @@ module.exports = {
           widgetPreview,
           contextualWidgetDefaultData,
           widgetManagers,
-          action: self.action,
-          widgetOperations
+          action: self.action
         };
       },
       async addDeduplicateWidgetIdsMigration() {